Remove obsolete map components and streamline main map view
- Deleted MapHeatmapComponent, MapObstaclesComponent, and MapRadarComponent to simplify the application structure.
- Updated MainComponent to consolidate map views into a single MapComponent for improved usability.
- Adjusted HTML and TypeScript files to reflect the removal of individual map components and enhance overall performance.

+++ Microcontroller/lib/Mapping/LiDAR/i2c_vl53l0x.h
@@ -17,7 +17,7 @@ bool i2c_read_addr8_data8(uint8_t , uint8_t *);
 bool i2c_read_addr8_data16(uint8_t , uint16_t *);
 bool i2c_read_addr16_data8(uint16_t , uint8_t *);
 bool i2c_read_addr16_data16(uint16_t , uint16_t *);
-bool i2c_read_addr8_data32(uint16_t , uint32_t *);
+bool i2c_read_addr8_data32(uint8_t , uint32_t *);
 bool i2c_read_addr16_data32(uint16_t , uint32_t *);
 
 /**
